Get calls calculate `used` value
Before this change GETs didn't reflect the fact that time has passed.
Now, when the server handles a GET request it calculates the updated
`used` value.

This change is to make GET calls consistent with their intent. The main
intent is that the developer can check how many items are used so maybe
show a status graph or something like that.

Error var's name is changed to ErrLimitReached more consistent with
other libraries and main go SDK.